requirements.txt
playwright==1.47.0
gspread==6.1.2
google-auth==2.34.0
pandas==2.2.2
python-dateutil==2.9.0
email-validator==2.2.0
dnspython==2.6.1
requests==2.32.3
scikit-learn==1.5.1
numpy==1.26.4
joblib==1.4.2
streamlit==1.38.0
plotly==5.24.0
